// src/styles/theme/theme.types.ts
export interface BreakpointsProps {
  maxWidth: number
  desktop: number
  tablet: number
  mobile: number
}

export interface FontsProps {
  primary: string
}

export interface MediaProps {
  smallMobile: string
  mobile: string
  tablet: string
  desktop: string
  maxWidth: string
}

export interface PaletteProps {
  blue: string
  lightblue: string
  beige: string
  lightbeige: string
  sandy: string
  ecru: string
  lightPink: string
  pink: string
  peach: string
  white: string
  lightgray: string
  black: string
}